Moves generate-zones-config to scripts folder
