+++ web/scripts/generate-zones-config.ts
